sysutils/beats: Refactor namespace
