+++ sysutils/beats/src/opnsense/mvc/app/controllers/OPNsense/Beats/forms/filebeat.xml
