+++ sysutils/beats/src/opnsense/mvc/app/models/OPNsense/Beats/Filebeat.xml
