--- sysutils/beats/src/opnsense/mvc/app/models/OPNsense/Beats8/ACL/ACL.xml
@@ -1,9 +1,0 @@
-<acl>
-    <page-services-beats8>
-        <name>Services: Beats8</name>
-        <patterns>
-            <pattern>ui/filebeat/*</pattern>
-            <pattern>api/filebeat/*</pattern>
-        </patterns>
-    </page-services-beats8>
-</acl>
