--- sysutils/beats/src/opnsense/mvc/app/models/OPNsense/Beats8/Menu/Menu.xml
@@ -1,7 +1,0 @@
-<menu>
-    <Services>
-        <Beats8 cssClass="fa fa-heartbeat fa-fw">
-            <Filebeat url="/ui/filebeat"/>
-        </Beats8>
-    </Services>
-</menu>
