+++ sysutils/beats/src/opnsense/service/conf/actions.d/actions_beats.conf
