+++ sysutils/beats/src/opnsense/service/templates/OPNsense/Beats/+TARGETS
