+++ sysutils/beats/src/opnsense/service/templates/OPNsense/Beats/filebeat.yml
